+++ dirty_cat/fast_hash.py
@@ -3,7 +3,7 @@ n-gram hashing by simple dot products
 
 The principle is as follows:
   1. A string is viewed as a succession of numbers (the ASCII or UTF8
-     representation of its elements.
+     representation of its elements).
   2. Each n-gram is then an n-dimensional vector of integers "g". A simple
      hash function is then computed by taking the dot product with a
      given random vector "atom", modulo max-int (integers larger than
@@ -25,20 +25,20 @@ MAXINT32 = np.int32(2 ** (32 - 1) - 1)
 
 @functools.lru_cache(maxsize=1024)
 def gen_atom(atom_len, seed=0):
-    """ Generate a random integer atom
+    """ Generate a random integer array (atom).
 
     Parameters
     ----------
-    atom_len: int
-        The length of the atom
-    seed: int, default=0
-        The seed of the random_number generator
+    atom_len : int
+        The length of the atom.
+    seed : int, default=0
+        The seed of the random_number generator.
 
     Returns
     -------
-    atom: 1D array of integers
+    array, shape (atom_len, )
         An array of random integers of length atom_len and dtype int32
-        (assuming dtype_size=32)
+        (assuming dtype_size=32).
     """
     rng = np.random.RandomState(seed)
     atom = rng.randint(-MAXINT32, MAXINT32, size=atom_len,
@@ -47,20 +47,25 @@ def gen_atom(atom_len, seed=0):
 
 
 def ngram_min_hash(string, ngram_range=(2, 4), seed=0, return_minmax=False):
-    """ Compute the min hash of the ngrams of the string
+    """ Compute the min/max hash of the ngrams of the string.
 
     Parameters
     ----------
-    string: str
-        String to min-hash
-    ngram_range: tuple (min_n, max_n)
+    string : str
+        String to encode.
+    ngram_range : tuple (min_n, max_n), default=(2, 4)
         The lower and upper boundary of the range of n-values 
         for different n-grams to be extracted.
-    seed: integer
-        Integer used to seed the hashing function
-    Return
+    seed : int, default=0
+        Integer used to seed the hashing function.
+    return_minmax : bool, default=False
+        If True, returns both the minhash and maxhash of the string.
+        Else, only returns the minhash. 
+    Returns
     -------
-    min_hash: integer
+    int or tuple 
+        The min_hash or (min_hash, max_hash) of the n-grams of the string.
+
     """
     # Create a numerical 1D array from the string
     array = np.frombuffer(string.encode(), dtype='int8', count=len(string))